requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.101.1
fastapi-users==12.1.1
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.5
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
PyJWT==2.8.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.18.2
uvloop==0.17.0
